simplify local webapp deployment
