--- TODO.txt
@@ -1,1 +1,0 @@
-TEST WITH AWK ON BIGGER DATA FOR EXTRACTION, SELECTION AND FILTERING (MAYBE ISSUE WITH TIME EXEC??)
